# spec/dashboard/games_openapi.yaml
openapi: 3.0.0
info:
  description: "Qilin api "
  version: "1.0.0"
  title: "Qilin games dashboard API"
  contact:
    email: "roman.golenok@protocol.one"
...
servers:
  - url: https://qilinapi.tst.protocol.one
    description: Master test server
paths:
...
  /api/v1/vendors/{id}/documents/reviews:
    delete:
      tags:
        - vendor
        - onboarding
      summary: "Revoke documents from review"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - vendor
        - onboarding
      summary: "Send documents about vendor to review"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: If-Unmodified-Since
          in: header
          schema:
            type: string
            format: date-time
          required: false
      responses:
        201:
          description: OK
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        409:
          $ref: '#/components/responses/Conflict'
        500:
          $ref: '#/components/responses/InternalError'
  /api/v1/vendors/{id}/documents:
    get:
      tags:
        - vendor
        - onboarding
      summary: "Get documents about vendor"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMerchantRequest'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'

    put:
      tags:
        - vendor
        - onboarding
      summary: "Change documents about vendor"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeMerchantRequest'
      responses:
        200:
          description: OK
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/GetMerchantRequest'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        409:
          $ref: '#/components/responses/Conflict'
          description: "Documents can be on review and have not to be changed"
        422:
          $ref: '#/components/responses/UnprocessableEntity'
        500:
          $ref: '#/components/responses/InternalError'
  /api/v1/vendors/{id}/messages/short:
    get:
      tags:
        - vendor
        - onboarding
        - notifications
      summary: "Get last notifications with short info"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShortNotification'
          headers:
            X-Centrifugo-Token:
              schema:
                type: string
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'
  /api/v1/vendors/{id}/messages:
    get:
      summary: "Get list of notifications"
      tags:
        - vendor
        - onboarding
        - notifications
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: "Limit for query"
          schema:
            type: integer
            default: 10
            minimum: 1
        - name: offset
          in: query
          description: "Offset for query"
          schema:
            type: integer
            default: 0
        - name: sort
          in: query
          description: "Notifications requests sorting"
          example: -createdAt,+title
          schema:
            type: "array"
            items:
              type: "string"
              enum:
                - -title
                - +title
                - -createdAt
                - +createdAt
                - -isUnread
                - +isUnread
        - name: query
          in: query
          example: "Error"
          description: "Notifications requests name filter"
          schema:
            type: string
      responses:
        200:
          description: OK
          headers:
            - name: X-Items-Count
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'
  /api/v1/vendors/{id}/messages/{messageId}/read:
    put:
      tags:
        - vendor
        - onboarding
        - notifications
      summary: "Sets status 'unread' to false"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalError'
  /api/v1/vendors/{id}/messages/{messageId}:
    get:
      tags:
        - vendor
        - onboarding
        - notifications
      summary: "Get full notification object"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal Server Error
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 409
            message: "Resource with same unique id already exists"
    UnprocessableEntity:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 422
            message: "Fields required: 'title, description'"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: 'date-time'
        isRead:
          type: boolean
    ShortNotification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        createdAt:
          type: string
          format: 'date-time'
        isRead:
          type: boolean
        haveMsg:
          type: boolean
...
    GetMerchantRequest:
      allOf:
        - $ref: '#/components/schemas/ChangeMerchantRequest'
        - type: object
          properties:
            status:
              type: string
              enum:
                - draft
                - on_review
                - approved
                - declined
    ChangeMerchantRequest:
      type: object
      properties:
        company:
          type: object
          properties:
            name:
              description: "company name"
              type: string
            alternativeName:
              type: string
              description: "Doing business as (alternative name)"
            country:
              type: string
              enum:
                - russia
                - germany
                - usa
                - etc.
              description: "Country (enum in lowercase)"
            region:
              description: "State / Province / Region"
              type: string
            zip:
              type: string
              pattern: "^[0-9]{5}(?:-[0-9]{4})?$"
              description: "Zip code."
            city:
              type: string
            address:
              type: string
            additionalAddress:
              type: string
              description: "Address line 2"
            registrationNumber:
              type: string
              example: "12394071075"
            taxId:
              type: string
              example: "4300098"
              description: "Tax ID / EIN / TIN"
        contact:
          type: object
          properties:
            authorized:
              type: object
              properties:
                fullName:
                  type: string
                  description: "Authorized person full name"
                position:
                  type: string
                  description: "Authorized person position"
                email:
                  type: string
                  format: email
                phone:
                  type: string
            technical:
              type: object
              properties:
                fullName:
                  type: string
                  description: "technical person full name"
                email:
                  type: string
                  format: email
                phone:
                  type: string
        banking:
          type: object
          properties:
            currency:
              type: string
              description: "payout currency"
              enum:
                - USD
                - EUR
                - RUR
            name:
              description: "Name of bank"
              type: string
              example: "Bank of Baroda"
            address:
              type: string
            accountNumber:
              type: string
            swift:
              type: string
            details:
              type: string
              description: "Corresponding bank details"
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
security:
  - bearerAuth: []